Updates libft.h and typing in functions

+++ ft_memcpy.c
@@ -6,11 +6,13 @@
 /*   By: lusantor <lusantor@student.42.rio>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/27 21:37:57 by lusantor          #+#    #+#             */
-/*   Updated: 2022/05/27 23:25:25 by lusantor         ###   ########.fr       */
+/*   Updated: 2022/06/09 17:34:49 by lusantor         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memcpy(void *restrict dest, const void *restrict src, int n) // int n should be typed size_t
+#include "libft.h"
+
+void	*ft_memcpy(void *restrict dest, const void *restrict src, size_t n)
 {
 	int i;
